optimization 1.0 proposal - cportfolio code base
