--- index.html
@@ -1,130 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto&display=swap" rel="stylesheet"> 
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <script src="https://cdn.jsdelivr.net/npm/handlebars@latest/dist/handlebars.js"></script>
-    
-    <script id="aside-template" type="text/x-handlebars-template">
-
-        {{#if links}}
-            {{#if (eq title "Work")}}
-                {{#each links}}
-                    <a id="{{hyperlink}}-link" href="#{{hyperlink}}" class="aside-link" onclick="handleWork(`{{hyperlink}}`)">
-                    {{hypertext}}</a>
-                {{/each}}
-           
-            {{else if (eq title "About")}}
-                    {{#each links}}
-                        <a id="{{hyperlink}}-link" href="#{{hyperlink}}" class="aside-link" onclick="handleAbout(`{{hyperlink}}`)">
-                        {{hypertext}}</a>
-                    {{/each}}
-        
-            {{else if (eq title "Contact")}}    
-                    {{#each links}}
-                        <a id="{{hyperlink}}-link" href="#{{hyperlink}}" class="aside-link" onclick="handleContact(`{{hyperlink}}`)">
-                        {{hypertext}}</a>
-                    {{/each}}  
-            {{/if}}
-        {{/if}}
-       
-    </script>
-    <title>cportfolio - a website template for career portfolio</title>
-</head>
-<body>
-    <!-- lets start our endeavor -->
-    <div class="container">
-        <header>
-            <h1 id="header-title">Hello!</h1>
-            <nav>
-                <a id="home" onclick="handleClick(this)" class="nav-link nav-active" onclick="handleClick(this)" href="#">Home</a>
-                <a id="work" onclick="handleClick(this)" class="nav-link" href="#">Work</a>
-                <a id="contact" onclick="handleClick(this)" class="nav-link" href="#">Contact</a>
-                <a id="about" onclick="handleClick(this)" class="nav-link" href="#">About</a>
-            </nav>
-        </header>
-        
-        <main>
-            <section class="content">
-
-                <!-- home folder -->
-                <div id="home-folder">
-                      <div class="home-text-wrapper">
-                          <h2>I'm John.</h2>
-                          <p>A tenacious, passionate marketer and maker who loves turning bold ideas into impactful & memorable experiences.</p>
-                      </div>                
-                </div>
-
-                <!-- work folder -->
-                <div id="work-folder">
-                    <div class="work-card" id="xydus-game">
-                        <img src="images/profile.jpg" alt="Xydus Game" />
-                        <h2>Xydus Game</h2>
-                        <p>Xydus is multi-player game</p>
-                        <p>Unreal</p>
-                    </div>
-                    <div class="work-card" id="xydus-music">
-                        <img src="images/profile.jpg" alt="Xydus Music" />
-                        <h2>Xydus Music</h2>
-                        <p>Xydus is online music streaming platform</p>
-                        <p>NodeJS</p>
-                    </div>
-                    <div class="work-card" id="poar-algo">
-                        <img src="images/profile.jpg" alt="Poar Algo" />
-                        <h2>Poar Algo</h2>
-                        <p>Poar Algo is a crypto currency</p>
-                        <p>Blockchain</p>
-                    </div>
-                </div>
-
-                <!-- contact folder -->
-                <div id="contact-folder">
-                    <div id="write-here" class="contact-card">
-                        Write here</div>
-                    <div id="email-list" class="contact-card">
-                        Email List</div>
-                    <div id="social-links" class="contact-card">
-                        <p>
-                            <a href="#"><i class="fa fa-dribbble"></i></a>
-                            <a href="#"><i class="fa fa-twitter"></i></a>
-                            <a href="#"><i class="fa fa-linkedin"></i></a>
-                            <a href="#"><i class="fa fa-facebook"></i></a>     
-                        </p>
-                    </div>
-                </div>
-
-                <!-- about folder -->
-                <div id="about-folder">
-                    <div id="profile" class="about-card">
-                        <img class="profile-image" src="images/profile.jpg" alt="John Doe" />
-                        <h2>John Doe</h2>
-                        <p>Designation</p>
-                        <p>Job title</p>
-                    </div> 
-                    <div id="skills" class="about-information">
-                        <p>Curious to know?</p>
-                    </div>
-                    <div id="education" class="about-information">
-                        <p>Education.</p>
-                    </div>
-                    <div id="bucket-list" class="about-information">
-                        <p>Bucket List</p>
-                    </div>
-
-                </div>
-            </section>
-    
-            <!-- aside -->
-            <aside id="aside"></aside>
-        </main>
-    </div>
-    <!-- <script src="js/index.js"></script> -->
-    <script src="js/site.config.js"></script>
-</body>
-</html>
